Fix judges page arrow bug
On clicking any of the arrows in the judges info page, the list of
speakers was being cycled through. This was fixed by adding a type
property to each speaker/judge object, which stores "speaker" or "judge"
depending on what kind of object it is. Henceforth the event handlers of
the arrows were updated to check the type and determine the
next/previous person on the basis of the said type